Rename cmd/zserio to cmd/go-zserio
This fixes "go install" to install the command as go-zserio.
